+++ src/main/java/org/apache/tomcat/jakartaee/ManifestConverter.java
@@ -33,19 +33,25 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.io.IOUtils;
+import org.eclipse.osgi.util.ManifestElement;
+import org.osgi.framework.BundleException;
+import org.osgi.framework.Constants;
 
 /**
  * Updates Manifests.
  */
 public class ManifestConverter implements Converter {
 
+    private static final String JAKARTA_SERVLET = "jakarta.servlet";
+    private static final Pattern SERVLET_PATTERN = Pattern.compile("jakarta.servlet([^,]*);version=\"(.*?)\"");
     private static final Logger logger = Logger.getLogger(ManifestConverter.class.getCanonicalName());
     private static final StringManager sm = StringManager.getManager(ManifestConverter.class);
 
     /**
      * Manifest converter constructor.
      */
-    public ManifestConverter() {}
+    public ManifestConverter() {
+    }
 
     @Override
     public boolean accepts(String filename) {
@@ -63,7 +69,8 @@ public class ManifestConverter implements Converter {
         Manifest srcManifest = new Manifest(new ByteArrayInputStream(srcBytes));
         Manifest destManifest = new Manifest(srcManifest);
 
-        // Only consider profile conversions, allowing Migration.hasConverted to be true only when there are actual
+        // Only consider profile conversions, allowing Migration.hasConverted to be true
+        // only when there are actual
         // conversions made
         boolean converted = updateValues(destManifest, profile);
         removeSignatures(destManifest);
@@ -79,7 +86,6 @@ public class ManifestConverter implements Converter {
 
         return converted;
     }
-
 
     private void removeSignatures(Manifest manifest) {
         manifest.getMainAttributes().remove(Attributes.Name.SIGNATURE_VERSION);
@@ -98,7 +104,6 @@ public class ManifestConverter implements Converter {
         }
     }
 
-
     private boolean isCryptoSignatureEntry(Attributes attributes) {
         for (Object attributeKey : attributes.keySet()) {
             if (attributeKey.toString().endsWith("-Digest")) {
@@ -108,7 +113,6 @@ public class ManifestConverter implements Converter {
         return false;
     }
 
-
     private boolean updateValues(Manifest manifest, EESpecProfile profile) {
         boolean converted = updateValues(manifest.getMainAttributes(), profile);
         for (Attributes attributes : manifest.getEntries().values()) {
@@ -116,7 +120,6 @@ public class ManifestConverter implements Converter {
         }
         return converted;
     }
-
 
     private boolean updateValues(Attributes attributes, EESpecProfile profile) {
         boolean converted = false;
@@ -128,11 +131,27 @@ public class ManifestConverter implements Converter {
             // Purposefully avoid setting result
         }
         // Update package names in values
-        for (Entry<Object,Object> entry : attributes.entrySet()) {
+        for (Entry<Object, Object> entry : attributes.entrySet()) {
             String newValue = profile.convert((String) entry.getValue());
-            newValue = replaceVersion(newValue);
+            String header = entry.getKey().toString();
+            try {
+                // Need to be careful with OSGI headers.
+                // Specifically, Export-Package cannot specify a version range.
+                // There may be other weird things as well (like directives that have
+                // jakarta.servlet packages).
+                if (Constants.IMPORT_PACKAGE.equals(header)) {
+                    newValue = processImportPackage(newValue);
+                } else if (Constants.EXPORT_PACKAGE.equals(header)) {
+                    newValue = processExportPackage(newValue);
+                } else {
+                    newValue = replaceVersion(newValue);
+                }
+            } catch (BundleException e) {
+                newValue = replaceVersion(newValue, !Constants.EXPORT_PACKAGE.equals(header));
+            }
+
             // Object comparison is deliberate
-            if (newValue != entry.getValue()) {
+            if (!newValue.equals(entry.getValue())) {
                 entry.setValue(newValue);
                 converted = true;
             }
@@ -140,12 +159,46 @@ public class ManifestConverter implements Converter {
         return converted;
     }
 
+    private String processExportPackage(String value) throws BundleException {
+        return processOSGIHeader(value, Constants.EXPORT_PACKAGE, "5.0.0");
+    }
+
+    private String processImportPackage(String value) throws BundleException {
+        return processOSGIHeader(value, Constants.IMPORT_PACKAGE, "[5.0.0,7.0.0)");
+    }
+
+    private String processOSGIHeader(String value, String header, String replacement) throws BundleException {
+        List<String> packages = new ArrayList<>();
+        ManifestElement[] elements = ManifestElement.parseHeader(header, value);
+        for (ManifestElement element : elements) {
+            if (element.getValue().startsWith(JAKARTA_SERVLET)) {
+                String oldVersion = element.getAttribute(Constants.VERSION_ATTRIBUTE);
+                if (oldVersion != null) {
+                    packages.add(element.toString().replace(oldVersion, replacement));
+                } else {
+                    packages.add(element.toString());
+                }
+            } else {
+                packages.add(element.toString());
+            }
+        }
+        if (packages.isEmpty()) {
+            return value;
+        }
+        return String.join(",", packages);
+    }
+
     private String replaceVersion(String entryValue) {
-        if (entryValue.contains("jakarta.servlet")) {
+        return replaceVersion(entryValue, true);
+    }
+
+    private String replaceVersion(String entryValue, boolean range) {
+        if (entryValue.contains(JAKARTA_SERVLET)) {
             StringBuffer builder = new StringBuffer();
-            Matcher matcher = Pattern.compile("jakarta.servlet([^,]*);version=\"(.*?)\"").matcher(entryValue);
+            Matcher matcher = SERVLET_PATTERN.matcher(entryValue);
             while (matcher.find()) {
-                matcher.appendReplacement(builder, "jakarta.servlet$1;version=\"[5.0.0,7.0.0)\"");
+                String version = range ? "[5.0.0,7.0.0)" : "5.0.0";
+                matcher.appendReplacement(builder, "jakarta.servlet$1;version=\"" + version + "\"");
             }
             matcher.appendTail(builder);
             return builder.toString();